refactor: split large files into smaller modules (Part 1)
- Reduced jira-client-service.ts from 1569 to 1260 lines by extracting issue operations
- Split jira-client.ts (1543 lines) into 8 focused modules under 700 lines each
- Split confluence-client-service.ts (1318 lines) into 11 focused modules under 700 lines each
- Updated precommit hook to set max file size threshold at 700 lines
- All extracted modules maintain full functionality and Effect patterns
- TypeScript compilation passes successfully

Remaining work:
- Several files still over 700 lines need further splitting
- This is Part 1 of the file size reduction effort
